requirements.txt
akshare==1.2.57
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==4.3.0
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.8
cryptography==36.0.0
cycler==0.11.0
decorator==5.1.0
distlib==0.3.1
Django==3.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.0.12
fonttools==4.28.3
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
jsonpath==0.82
kiwisolver==1.3.2
kombu==4.6.11
lxml==4.6.4
matplotlib==3.5.0
numpy==1.21.4
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
py-mini-racer==0.6.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.6
pypinyin==0.44.0
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
setuptools-scm==6.3.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tabulate==0.8.9
tomli==1.2.2
tqdm==4.62.3
urllib3==1.26.7
vine==1.3.0
virtualenv==20.4.3
webencodings==0.5.1
websocket-client==1.2.1
xlrd==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
